move tests out of rome_app - divide them into unit_tests that can be run locally and by CI, and integration_tests that need access to the dev server
