+++ run_integration_tests.py
@@ -1,0 +1,15 @@
+#!/usr/bin/env python
+import os
+import sys 
+import django
+from django.conf import settings
+from django.test.utils import get_runner
+
+
+if __name__ == '__main__':
+    os.environ['DJANGO_SETTINGS_MODULE'] = 'config.settings'
+    django.setup()
+    TestRunner = get_runner(settings)
+    test_runner = TestRunner()
+    failures = test_runner.run_tests(['integration_tests'])
+    sys.exit(bool(failures))
