+++ unit_tests/tests.py
@@ -2,8 +2,8 @@
 from django.contrib.auth.models import User
 from django.core.urlresolvers import reverse
 from django.test import TestCase, TransactionTestCase, Client
-from . import views
-from . import models
+from rome_app import views
+from rome_app import models
 
 
 def get_auth_client():
@@ -43,20 +43,6 @@ class TestStaticViews(TestCase):
 
 class TestBooksViews(TestCase):
 
-    def test_books(self):
-        response = self.client.get(reverse('books'))
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, 'Rome - Books')
-
-    def test_book_thumnbail_viewer(self):
-        response = self.client.get(reverse('thumbnail_viewer', kwargs={'book_id': '230605'}))
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, 'Book 230605')
-
-    def test_book_page_viewer(self):
-        response = self.client.get(reverse('book_page_viewer', kwargs={'book_id': '230605', 'page_id': '230606'}))
-        self.assertEqual(response.status_code, 200)
-
     def test_new_annotation_auth(self):
         url = reverse('new_annotation', kwargs={'book_id': '230605', 'page_id': '230606'})
         response = self.client.get(url)
@@ -73,13 +59,6 @@ class TestBooksViews(TestCase):
         url = reverse('edit_annotation', kwargs={'book_id': '224807', 'page_id': '224895', 'anno_id': '228874'})
         response = self.client.get(url)
         self.assertRedirects(response, 'http://testserver/rome/login/?next=%s' % url)
-
-    def test_edit_annotation_get(self):
-        auth_client = get_auth_client()
-        url = reverse('edit_annotation', kwargs={'book_id': '224807', 'page_id': '224895', 'anno_id': '228874'})
-        response = auth_client.get(url)
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, 'value="Submit Annotation"')
 
     def test_get_next_prev_pids(self):
         prev_id, next_id = views._get_prev_next_ids({'relations': {'hasPart': []}}, None)
@@ -105,15 +84,6 @@ class TestBooksViews(TestCase):
 
 class TestPrintsViews(TestCase):
 
-    def test_prints(self):
-        response = self.client.get(reverse('prints'))
-        self.assertEqual(response.status_code, 200)
-
-    def test_specific_print(self):
-        response = self.client.get(reverse('specific_print', kwargs={'print_id': 230631}))
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, 'ARTES ATHENIX PARTHENOPEM')
-
     def test_new_print_annotation_auth(self):
         url = reverse('new_print_annotation', kwargs={'print_id': '230631'})
         response = self.client.get(url)
@@ -130,13 +100,6 @@ class TestPrintsViews(TestCase):
         url = reverse('edit_print_annotation', kwargs={'print_id': '230631', 'anno_id': '230632'})
         response = self.client.get(url)
         self.assertRedirects(response, 'http://testserver/rome/login/?next=%s' % url)
-
-    def test_edit_print_annotation_get(self):
-        auth_client = get_auth_client()
-        url = reverse('edit_print_annotation', kwargs={'print_id': '230631', 'anno_id': '230632'})
-        response = auth_client.get(url)
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, 'value="Submit Annotation"')
 
 
 class TestEssaysViews(TestCase):
@@ -163,12 +126,6 @@ class TestPeopleViews(TransactionTestCase):
         response = self.client.get(reverse('people'))
         self.assertEqual(response.status_code, 200)
         self.assertContains(response, u'Frëd')
-
-    def test_person_detail(self):
-        models.Biography.objects.create(name=u'Frëd', trp_id='0001', bio=u'### Frëd')
-        response = self.client.get(reverse('person_detail', kwargs={'trp_id': '0001'}))
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, u'<h3>Frëd</h3>')
 
 
 class TestRecordCreatorViews(TestCase):
